Integrated Axes (Axis2D / Axis3D)

Note

Axis2D refers to the old Makie workflow in which the Scene was the main object for users to interact with. Back then, the default 2D axis was created like a plot object and therefore had many usability issues that were resolved with the new layout system and the new Axis type. Currently, the 3D axis still uses the old workflow, although you usually embed a scene with a 3D axis in a layout using the LScene wrapper.

The Axis2D or Axis3D is just a scene object, making it easy to manipulate and share between plots. These objects also contain the mapping you want to apply to the data and can interactively be changed. They can be created from any boundingbox and inserted into any plot.

There are two types of plot-like axes: Axis2D and Axis3D.

Interacting with the integrated axis

One can quite easily interact with the attributes of the axis like with any other plot.

You can access the axis of a scene by doing

axis = scene[OldAxis]

The axis attributes are nested, and there are different attributes depending on whether it is an Axis2D or Axis3D object.

You can access the nested attributes in multiple ways. Take the nested attribute axis -> :names -> :axisnames, for example:

axis[:names, :axisnames] = ("x", "y", "z")

axis[:names][:axisnames] = ("x", "y", "z")

axis = (names = (axisnames = ("x", "y", "z"),),)

Convenience functions for integrated axes

Makie offers some convenience functions to make manipulating the Axis2D / Axis3D easier.

Makie.xlims! — Function

xlims!(limits::Real...)
xlims!(limits::NTuple{2, Real})
xlims!(scene, limits::Real...)
xlims!(scene, limits::NTuple{2, Real})

Set the x-limits for the given Scene (defaults to current Scene).

Makie.ylims! — Function

ylims!(limits::Real...)
ylims!(limits::NTuple{2, Real})
ylims!(scene, limits::Real...)
ylims!(scene, limits::NTuple{2, Real})

Set the y-limits for the given Scene (defaults to current Scene).

Makie.zlims! — Function

zlims!(limits::Real...)
zlims!(limits::NTuple{2, Real})
zlims!(scene, limits::Real...)
zlims!(scene, limits::NTuple{2, Real})

Set the z-limits for the given Scene (defaults to current Scene).

Makie.xlabel! — Function

xlabel!([scene,] xlabel)

Set the x-axis label for the given Scene. Defaults to using the current Scene.

Makie.ylabel! — Function

ylabel!([scene,] ylabel)

Set the y-axis label for the given Scene. Defaults to using the current Scene.

Makie.zlabel! — Function

zlabel!([scene,] zlabel)

Set the z-axis label for the given Scene. Defaults to using the current Scene.

Warning

The Scene must have an Axis3D. If not, then this function will error.

Makie.xticklabels — Function

xticklabels(scene)

Returns the all the x-axis tick labels. See also ticklabels.

Makie.yticklabels — Function

yticklabels(scene)

Returns the all the y-axis tick labels. See also ticklabels.

Makie.zticklabels — Function

zticklabels(scene)

Returns the all the z-axis tick labels. See also ticklabels.

Makie.xtickrange — Function

xtickrange(scene)

Returns the tick range along the x-axis. See also tickranges.

Makie.ytickrange — Function

ytickrange(scene)

Returns the tick range along the y-axis. See also tickranges.

Makie.ztickrange — Function

ztickrange(scene)

Returns the tick range along the z-axis. See also tickranges.

Makie.xticks! — Function

xticks!([scene,]; xtickrange=xtickrange(scene), xticklabels=xticklabel(scene))

Set the tick labels and range along the x-axes. See also ticks!.

Makie.yticks! — Function

yticks!([scene,]; ytickrange=ytickrange(scene), yticklabels=yticklabel(scene))

Set the tick labels and range along all the y-axis. See also ticks!.

Makie.zticks! — Function

zticks!([scene,]; ztickranges=ztickrange(scene), zticklabels=zticklabel(scene))

Set the tick labels and range along all z-axis. See also ticks!.

Makie.xtickrotation — Function

xtickrotation(scene)

Returns the rotation of tick labels along the x-axis. See also tickrotations

Makie.ytickrotation — Function

ytickrotation(scene)

Returns the rotation of tick labels along the y-axis. See also tickrotations

Makie.ztickrotation — Function

ztickrotation(scene)

Returns the rotation of tick labels along the z-axis. See also tickrotations

Makie.xtickrotation! — Function

xtickrotation!([scene,] xangle)

Set the rotation of tick labels along the x-axis. See also tickrotations!.

Makie.ytickrotation! — Function

ytickrotation!([scene,] yangle)

Set the rotation of tick labels along the y-axis. See also tickrotations!.

Makie.ztickrotation! — Function

ztickrotation!([scene,] zangle)

Set the rotation of tick labels along the z-axis. See also tickrotations!.

Axis3D

Axis3D attributes groups

Symbol	Description

frame	See the detailed descriptions for frame attributes.

names	See the detailed descriptions for names attributes.

scale	NTuple{3,Float}. Specifies the scaling for the axes.

showaxis	NTuple{3,Bool}. Specifies whether to show the axes.

showgrid	NTuple{3,Bool}. Specifies whether to show the axis grids.

showticks	NTuple{3,Bool}. Specifies whether to show the axis ticks.

ticks	See the detailed descriptions for ticks attributes.

:ticks

Symbol	Description

align	NTuple{3,(:pos, :pos)}. Specify the text alignment for the axis ticks, where :pos can be :left, :center, or :right.

font	NTuple{3,String}. Specifies the font for the axis ticks, and can choose any font available on the system.

gap	Number. Specifies the gap (in pixels) between the axis ticks and the axes themselves.

rotation	NTuple{3,Quaternion{Float32}}. Specifies the rotations for each axis's ticks, in radians.

textcolor	NTuple{3,Symbol or Colorant}. Specifies the color of the axes ticks. Can be a color symbol/string like :red, or a Colorant.

textsize	Integer. Font pointsize for text.

:frame

Symbol	Description

axiscolor	Symbol or Colorant. Specifies the color of the axes. Can be a color symbol/string like :red, or a Colorant.

linecolor	Symbol or Colorant. Specifies the color of the grid lines. Can be a color symbol/string like :red, or a Colorant.

linewidth	Number. Width of the axes grid lines.

:names

Symbol	Description

align	NTuple{3,(:pos, :pos)}. Specify the text alignment for the axis labels, where :pos can be :left, :center, or :right.

axisnames	NTuple{3,String}. Specifies the axis labels.

font	NTuple{3,String}. Specifies the font for the axis labels, and can choose any font available on the system.

gap	Number. Specifies the gap (in pixels) between the axis labels and the axes themselves.

rotation	NTuple{3,Quaternion{Float32}}. Specifies the rotations for each axis's label, in radians.

textcolor	NTuple{3,Symbol or Colorant}. Specifies the color of the axes labels. Can be a color symbol/string like :red, or a Colorant.

textsize	NTuple{3,Int}. Font pointsize for axes labels.