Help functions

Help functions

help

AbstractPlotting.help — Function.

help(func[; extended = false])

Welcome to the main help function of Makie.jl / AbstractArray.jl.

For help on a specific function's arguments, type help_arguments(function_name).

For help on a specific function's attributes, type help_attributes(plot_Type).

Use the optional extended = true keyword argument to see more details.

source

Example usage:

help(scatter; extended = true)

```
`scatter(x, y, z)` / `scatter(x, y)` / `scatter(positions)`
```

Plots a marker for each element in `(x, y, z)`, `(x, y)`, or `positions`.

`AbstractPlotting.scatter` has the following function signatures:

```
  (Vector, Vector)
  (Vector, Vector, Vector)
  (Matrix)
```

Please refer to [`convert_arguments`](@ref) to find the full list of accepted arguments

Available attributes and their defaults for `Scatter{...} where T` are:

```
  alpha             1.0
  color             :black
  colormap          :viridis
  colorrange        AbstractPlotting.Automatic()
  glowcolor         RGBA{N0f8}(0.0,0.0,0.0,0.0)
  glowwidth         0.0
  linewidth         1
  marker            GeometryTypes.HyperSphere{2,T} where T
  marker_offset     AbstractPlotting.Automatic()
  markersize        0.1
  rotations         AbstractPlotting.Billboard()
  strokecolor       RGBA{N0f8}(0.0,0.0,0.0,0.0)
  strokewidth       0.0
  transform_marker  false
  visible           true
```

You can see usage examples of `AbstractPlotting.scatter` by running:

`example_database(AbstractPlotting.scatter)`

help_arguments

AbstractPlotting.help_arguments — Function.

help_arguments(io, func)

Returns a list of signatures for function func.

source

Example usage:

help_arguments(STDOUT, scatter)

`AbstractPlotting.scatter` has the following function signatures:

```
  (Vector, Vector)
  (Vector, Vector, Vector)
  (Matrix)
```

help_attributes

AbstractPlotting.help_attributes — Function.

help_attributes(io, Typ[; extended = false])

Returns a list of attributes for the plot type Typ. The attributes returned extend those attributes found in the default_theme.

Use the optional keyword argument extended (default = false) to show in addition the default values of each attribute.

source

Example usage:

help_attributes(STDOUT, Scatter; extended = true)

Available attributes and their defaults for `Scatter{...} where ArgType` are:

```
  alpha             1.0
  color             :black
  colormap          :viridis
  colorrange        AbstractPlotting.Automatic()
  glowcolor         RGBA{N0f8}(0.0,0.0,0.0,0.0)
  glowwidth         0.0
  linewidth         1
  marker            GeometryTypes.HyperSphere{2,T} where T
  marker_offset     AbstractPlotting.Automatic()
  markersize        0.1
  rotations         AbstractPlotting.Billboard()
  strokecolor       RGBA{N0f8}(0.0,0.0,0.0,0.0)
  strokewidth       0.0
  transform_marker  false
  visible           true
```

Plot styling options

Use these functions to find out the styling options.

available_marker_symbols

AbstractPlotting.available_marker_symbols()

Marker Symbols:
    + => +
    diamond => ◆
    star4 => ✦
    star8 => ✷
    star6 => 🟋
    rtriangle => ▶
    x => x
    circle => ●
    pentagon => ⬟
    vline => ┃
    cross => ✚
    xcross => ❌
    rect => ■
    ltriangle => ◀
    dtriangle => ▼
    utriangle => ▲
    star5 => ★
    octagon => ⯄
    hline => ━
    hexagon => ⬢

available_gradients

using AbstractPlotting # hide
AbstractPlotting.available_gradients()

For other plot attributes and their usage, see the section Plot attributes.