Output

Output

Makie overloads the FileIO interface, so it is simple to save plots as images or videos.

Static plots

To save a scene as an image, you can just write e.g.:

Makie.save("plot.png", scene)
Makie.save("plot.jpg", scene)

where scene is the scene handle.

In the backend, ImageMagick is used for the image format conversions.

Animated plots

It is also possible to output animated plots as videos (note that this requires ffmpeg to be installed and properly configured on your computer (test this by running ffmpeg -version from a terminal window).)

Makie.record — Function.

record(func, scene, path)

usage:

    record(scene, "test.gif") do io
        for i = 1:100
            scene.plots[:color] = ...# animate scene
            recordframe!(io) # record a new frame
        end
    end

source

record(func, scene, path, iter)

usage:

    record(scene, "test.gif", 1:100) do i
        scene.plots[:color] = ...# animate scene
    end

source

For recording of videos (either as .mp4 or .gif), you can do:

record(scene, "video.mp4", itr) do i
    func(i) # or some other animation in scene
end

where itr is an iterator and scene is the scene handle.

It is also possible to record to gifs:

record(scene, "video.gif", itr) do i
    func(i) # or some other animation in scene
end

In both cases, the returned value is a path pointing to the location of the recorded file.

Example usage

using Makie

scene = Scene()

f(t, v, s) = (sin(v + t) * s, cos(v + t) * s, (cos(v + t) + sin(v)) * s)
t = Node(Base.time()) # create a life signal
limits = FRect3D(Vec3f0(-1.5, -1.5, -3), Vec3f0(3, 3, 6))
p1 = meshscatter!(scene, lift(t-> f.(t, linspace(0, 2pi, 50), 1), t), markersize = 0.5)[end]
p2 = meshscatter!(scene, lift(t-> f.(t * 2.0, linspace(0, 2pi, 50), 1.5), t), markersize = 0.5)[end]

lines = lift(p1[1], p2[1]) do pos1, pos2
    map((a, b)-> (a, b), pos1, pos2)
end
linesegments!(scene, lines, linestyle = :dot, limits = limits)
# record a video
record(scene, "/home/s/.julia/v0.6/Makie/docs/media/videostream.mp4", 1:300) do i
    push!(t, Base.time())
end

For more info, consult the Examples gallery.