Conversions

Conversions

Every attribute has a conversion function, allowing attributes to take in rich input types, but keep the variance low for the backends. The conversions are overloadable as explained in Extending, making it simple to integrate custom types.

MakiE.to_color — Function.

to_color(b, x)

Colors.Colorants

source

A Symbol naming a color, e.g. :black

source

A String naming a color, e.g. :black or html style #rrggbb

source

A Tuple or Array with elements that to_color accepts. If Array is a Matrix it will get interpreted as an Image

source

Tuple{<: ColorLike, <: AbstractFloat} for a transparent color

source

MakiE.to_colormap — Function.

to_colormap(b, x)

An AbstractVector{T} with any object that to_color accepts

source

Tuple(A, B) or Pair{A, B} with any object that to_color accepts

source

A Symbol/String naming the gradient. For more on what names are available please see: `available_gradients()

source

MakiE.to_colornorm — Function.

to_colornorm(b, norm, intensity)

anything that can be converted to Vec2f0 (e.g. Tuple, Vector)

source

If colornorm is nothing will default to calculate the extrema from intensity

source

MakiE.to_scale — Function.

to_scale(b, s::Number)::Vec

source

to_scale(b, s::VecLike)::Point

source

MakiE.to_offset — Function.

to_offset(b, s::Number)::Point

source

to_scale(b, s::VecLike)::Point

source

MakiE.to_rotation — Function.

to_rotation(b, vec4)

source

to_rotation(b, quaternion)

source

to_rotation(b, tuple_float)

source

MakiE.to_image — Function.

to_image(b, image)

All kinds of images

source

MakiE.to_bool — Function.

to_bool(b, bool)

To boolean

source

to_bool(b, tuple)

Accepts Tuple of Bool

source

MakiE.to_index_buffer — Function.

to_index_buffer(b, x::GLBuffer{UInt32})

source

TOrSignal{Int}, AbstractVector{UnitRange{Int}}, TOrSignal{UnitRange{Int}}

source

AbstractVector{<:Integer} assumend 1-based indexing

source

AbstractVector{<:Face{2}} for linesegments

source

AbstractVector{UInt32}, is assumed to be 0 based

source

MakiE.to_positions — Function.

to_positions(b, positionlike)

NTuple{2, AbstractArray{Float}} for 2D points

source

NTuple{3, AbstractArray{Float}} for 3D points

source

view(AbstractArray{Point}, idx) for a subset of points. Can be shared (so you can plot subsets of the same data)!

source

AbstractArray{T} where T needs to have length defined and must be convertible to a Point

source

MakiE.to_array — Function.

to_array(b, arraylike)

AbstractArray

source

MakiE.to_scalefunc — Function.

to_scalefunc(b, x)

Function

source

MakiE.to_text — Function.

to_text(b, x)

All text

source

MakiE.to_font — Function.

to_font(b, x)

All fonts

source

MakiE.to_intensity — Function.

AbstractArray

source

MakiE.to_surface — Function.

to_surface(b, x::Range)

Range

source

to_surface(b, arraylike)

Anything that can be converted to Matrix/Vector of Float32

source

MakiE.to_spritemarker — Function.

to_spritemarker(b, x::Circle)

GeometryTypes.Circle(Point2(...), radius)

source

to_spritemarker(b, ::Type{Circle})

Type{GeometryTypes.Circle}

source

to_spritemarker(b, ::Type{Rectangle})

Type{GeometryTypes.Rectangle}

source

to_spritemarker(b, marker::Char)

Any Char, including unicode

source

Matrix of AbstractFloat will be interpreted as a distancefield (negative numbers outside shape, positive inside)

source

Any AbstractMatrix{<: Colorant} or other image type

source

A Symbol - Available options can be printed with available_marker_symbols()

source

Vector of anything that is accepted as a single marker will give each point it's own marker. Note that it needs to be a uniform vector with the same element type!

source

MakiE.to_static_vec — Function.

to_static_vec(b, x)

AbstractArray

source

Any StaticVector

source

NTuple

source

Abstract array of which the elements can be converted to vec

source

MakiE.to_rotations — Function.

to_rotations(b, x)

Billboard() for a rotation that will always face the camera

source

Any AbstractArray which elements can be converted to Vec4 (as a quaternion x, y, z, w)

source

MakiE.to_markersize2d — Function.

to_markersize2d(b, x)

Anything that can be converted to Vec2f0 for x, y scale

source

MakiE.to_markersize3d — Function.

to_markersize3d(b, x)

Anything that can be converted to Vec3f0 for x, y, z scale

source

MakiE.to_linestyle — Function.

to_linestyle(b, x)

Nothing for no style

source

AbstractVector{<:AbstractFloat} for denoting sequences of fill/nofill. E.g. [0.5, 0.8, 1.2] will result in 0.5 filled, 0.3 unfilled, 0.4 filled. 1.0 unit is one linewidth!

source

A Symbol equal to :dash, :dot, :dashdot, :dashdotdot

source

MakiE.to_normals — Function.

to_normals(b, x)

Vector{Normal{3}}

source

MakiE.to_faces — Function.

to_faces(b, x)

Any array of NTuple/GeometryTypes.Face

source

MakiE.to_attribut_id — Function.

to_attribut_id(b, x)

Index into Mesh attributes, Vector{Integer}

source

MakiE.to_mesh — Function.

to_mesh(b, meshlike)

AbstractMesh

source

MakiE.to_float — Function.

to_float(b, x)

Any Object convertible to Floatingpoint

source

MakiE.to_spatial_order — Function.

to_spatial_order(b, x)

"xy" or "yx"

source

:xy or :yx

source

MakiE.to_interval — Function.

to_interval(b, x)

Tuple{<: Number, <: Number}

source

Pair{<: Number, <: Number} e.g. 2 => 100

source

AbstractVector will be interpreted as an interval from minimum to maximum

source

MakiE.to_volume_algorithm — Function.

to_volume_algorithm(b, x)

Enum values: IsoValue Absorption MaximumIntensityProjection AbsorptionRGBA IndexedAbsorptionRGBA

source

Symbol/String: iso, absorption, mip, absorptionrgba, indexedabsorption

source

MakiE.to_3floats — Function.

3 Numbers for each dimension

source