Functions

Functions

Primitive plotting functions. These are the most atomic operations from which one can stack together more complex plots

MakiE.scatter — Function.

scatter(x, y, z) / scatter(x, y) / scatter(positions)

Plots a marker for each element in xyz/positions

Attributes:

Attribute x, convert function to_array which accepts:

`AbstractArray`

____________________

Attribute y, convert function to_array which accepts:

`AbstractArray`

____________________

Attribute positions, convert function to_positions which accepts:

`NTuple{2, AbstractArray{Float}}` for 2D points

`NTuple{3, AbstractArray{Float}}` for 3D points

`view(AbstractArray{Point}, idx)` for a subset of points. Can be shared (so you can plot subsets of the same data)!

`AbstractArray{T}` where T needs to have `length` defined and must be convertible to a Point

____________________

Attribute colormap, convert function to_colormap which accepts:

An `AbstractVector{T}` with any object that [to_color](@ref) accepts

Tuple(A, B) or Pair{A, B} with any object that [to_color](@ref) accepts

A Symbol/String naming the gradient. For more on what names are available please see: `available_gradients()`

____________________

Attribute intensity, convert function to_intensity which accepts:

`AbstractArray`

____________________

Attribute colornorm, convert function to_colornorm which accepts:

anything that can be converted to `Vec2f0` (e.g. `Tuple`, `Vector`)

If colornorm is `nothing` will default to calculate the extrema from `intensity`

____________________

Attribute marker, convert function to_spritemarker which accepts:

`GeometryTypes.Circle(Point2(...), radius)`

`Type{GeometryTypes.Circle}`

`Type{GeometryTypes.Rectangle}`

Any `Char`, including unicode

Matrix of AbstractFloat will be interpreted as a distancefield (negative numbers outside shape, positive inside)

Any AbstractMatrix{<: Colorant} or other image type

A `Symbol` - Available options can be printed with `available_marker_symbols()`

Vector of anything that is accepted as a single marker will give each point it's own marker. Note that it needs to be a uniform vector with the same element type!

____________________

Attribute strokecolor, convert function to_color which accepts:

`Colors.Colorants`

A `Symbol` naming a color, e.g. `:black`

A `String` naming a color, e.g. `:black` or html style `#rrggbb`

A Tuple or Array with elements that `to_color` accepts

____________________

Attribute strokewidth, convert function to_float which accepts:

Any Object convertible to Floatingpoint

____________________

Attribute glowcolor, convert function to_color which accepts:

`Colors.Colorants`

A `Symbol` naming a color, e.g. `:black`

A `String` naming a color, e.g. `:black` or html style `#rrggbb`

A Tuple or Array with elements that `to_color` accepts

____________________

Attribute glowwidth, convert function to_float which accepts:

Any Object convertible to Floatingpoint

____________________

Attribute markersize, convert function to_markersize which accepts:

Anything that can be converted to `Vec2f0` for x, y scale

____________________

Attribute rotations, convert function to_rotations which accepts:

`Billboard()` for a rotation that will always face the camera

Any AbstractArray which elements can be converted to Vec4 (as a quaternion x, y, z, w)

____________________

source

MakiE.lines — Function.

lines(x, y, z) / lines(x, y) / lines(positions)

Plots a connected line for each element in xyz/positions

Attributes:

Attribute x, convert function to_array which accepts:

`AbstractArray`

____________________

Attribute y, convert function to_array which accepts:

`AbstractArray`

____________________

Attribute positions, convert function to_positions which accepts:

`NTuple{2, AbstractArray{Float}}` for 2D points

`NTuple{3, AbstractArray{Float}}` for 3D points

`view(AbstractArray{Point}, idx)` for a subset of points. Can be shared (so you can plot subsets of the same data)!

`AbstractArray{T}` where T needs to have `length` defined and must be convertible to a Point

____________________

Attribute colormap, convert function to_colormap which accepts:

An `AbstractVector{T}` with any object that [to_color](@ref) accepts

Tuple(A, B) or Pair{A, B} with any object that [to_color](@ref) accepts

A Symbol/String naming the gradient. For more on what names are available please see: `available_gradients()`

____________________

Attribute intensity, convert function to_intensity which accepts:

`AbstractArray`

____________________

Attribute colornorm, convert function to_colornorm which accepts:

anything that can be converted to `Vec2f0` (e.g. `Tuple`, `Vector`)

If colornorm is `nothing` will default to calculate the extrema from `intensity`

____________________

Attribute linewidth, convert function to_float which accepts:

Any Object convertible to Floatingpoint

____________________

Attribute linestyle, convert function to_linestyle which accepts:

`Nothing` for no style

`AbstractVector{<:AbstractFloat}` for denoting sequences of fill/nofill. E.g. [0.5, 0.8, 1.2] will result in 0.5 filled, 0.3 unfilled, 0.4 filled

A `Symbol` equal to `:dash`, `:dot`, `:dashdot`, `:dashdotdot`

____________________

Attribute pattern, convert function to_pattern which accepts:

Same as `to_linestyle`

____________________

source

MakiE.surface — Function.

surface(x, y, z)

Plots a surface, where x y z are supposed to lie on a grid

Attributes:

Attribute x, convert function to_surface which accepts:

`Range`

Anything that can be converted to Matrix/Vector of Float32

____________________

Attribute y, convert function to_surface which accepts:

`Range`

Anything that can be converted to Matrix/Vector of Float32

____________________

Attribute z, convert function to_surface which accepts:

`Range`

Anything that can be converted to Matrix/Vector of Float32

____________________

Attribute color, convert function to_color which accepts:

`Colors.Colorants`

A `Symbol` naming a color, e.g. `:black`

A `String` naming a color, e.g. `:black` or html style `#rrggbb`

A Tuple or Array with elements that `to_color` accepts

____________________

source

MakiE.mesh — Function.

mesh(x, y, z) / mesh(mesh_object)

Plots a 3D mesh

Attributes:

Attribute color, convert function to_color which accepts:

`Colors.Colorants`

A `Symbol` naming a color, e.g. `:black`

A `String` naming a color, e.g. `:black` or html style `#rrggbb`

A Tuple or Array with elements that `to_color` accepts

____________________

Attribute shading, convert function Bool which accepts:

```
Bool <: Integer
```

Boolean type.

____________________

Attribute positions, convert function to_positions which accepts:

`NTuple{2, AbstractArray{Float}}` for 2D points

`NTuple{3, AbstractArray{Float}}` for 3D points

`view(AbstractArray{Point}, idx)` for a subset of points. Can be shared (so you can plot subsets of the same data)!

`AbstractArray{T}` where T needs to have `length` defined and must be convertible to a Point

____________________

Attribute faces, convert function to_faces which accepts:

Any array of NTuple/GeometryTypes.Face

____________________

Attribute attribute_id, convert function to_attribut_id which accepts:

Index into Mesh attributes, Vector{Integer}

____________________

Attribute vertexmesh, convert function to_mesh which accepts:

X, Y, Z as triangles, so need to have the same length and be dividable by 3

____________________

source

MakiE.wireframe — Function.

wireframe(x, y, z) / wireframe(positions) / wireframe(mesh)

Draws a wireframe either interpreted as a surface or mesh

Attributes:

source

image heatmap volume text poly